Struct mozangle::shaders::ShaderValidator

pub struct ShaderValidator {
    handle: ShHandle,
}

Fields
handle: ShHandle
Implementations

impl ShaderValidator

pub fn new( shader_type: u32, spec: ShaderSpec, output: Output, resources: &BuiltInResources, ) -> Option<ShaderValidator>

Create a new ShaderValidator instance NB: To call this you should have called first initialize()

pub fn for_webgl( shader_type: u32, output: Output, resources: &BuiltInResources, ) -> Option<ShaderValidator>

pub fn for_webgl2( shader_type: u32, output: Output, resources: &BuiltInResources, ) -> Option<ShaderValidator>

pub fn compile( &self, strings: &[&str], options: ShCompileOptions, ) -> Result<(), &'static str>

pub fn object_code(&self) -> String

pub fn info_log(&self) -> String

pub fn compile_and_translate( &self, strings: &[&str], ) -> Result<String, &'static str>

pub fn uniform_name_map(&self) -> HashMap<String, String>

Returns a map from uniform name in the original shader to uniform name in the compiled shader.

The map can be empty if the SH_VARIABLES option wasn’t specified.

pub fn get_num_unpacked_varying_vectors(&self) -> i32